requirements.txt
Flask
Flask-CORS
Werkzeug
supabase
python-dotenv